test(webapp): split dashboard watch suites
