+++ apps/webapp/test/dashboardAgentWatches.delivery.test.ts
@@ -1,0 +1,809 @@
+import {
+  appendChatMessageOnce,
+  cancelWatch,
+  chatExists,
+  claimWatchDelivery,
+  claimWatchTick,
+  countUnreadWatchWakes,
+  getChatMessages,
+  getWatch,
+  listActiveWatchesForChat,
+  listChatIdsWithUnreadWakes,
+  listRecentWatchWakes,
+  markWatchDelivered,
+  readWatchWakeFeed,
+  recordWatchCheck,
+  releaseWatchDelivery,
+  transitionWatchCondition,
+  WATCH_DELIVERY_CLAIM_STALE_MS,
+  type DashboardAgentDb,
+  type Watch,
+} from "@internal/dashboard-agent-db";
+import { postgresTest } from "@internal/testcontainers";
+import type { PrismaClient } from "@trigger.dev/database";
+import { afterEach, beforeEach, describe, expect, vi } from "vitest";
+import type { WatchCheckDeps } from "~/services/dashboardAgentWatchChecks";
+import {
+  DashboardAgentWatchesTestHarness,
+  RUN_START,
+  type DashboardAgentWatchesTestContext,
+  type Seeded,
+} from "./helpers/dashboardAgentWatchesTestHelpers";
+
+vi.setConfig({ testTimeout: 60_000 });
+
+const ctx = vi.hoisted(
+  (): DashboardAgentWatchesTestContext => ({
+    prisma: undefined as unknown as PrismaClient,
+    agentDb: undefined as unknown as DashboardAgentDb,
+    canAccess: true,
+    actor: undefined,
+    triggered: [],
+  })
+);
+
+vi.mock("~/db.server", () => {
+  const proxy = new Proxy(
+    {},
+    { get: (_target, prop) => (ctx.prisma as unknown as Record<string, unknown>)[prop as string] }
+  );
+  return { prisma: proxy, $replica: proxy, sqlDatabaseSchema: undefined };
+});
+
+vi.mock("~/services/dashboardAgentDb.server", () => ({
+  get dashboardAgentDb() {
+    return ctx.agentDb;
+  },
+}));
+
+process.env.SESSION_SECRET = "test-session-secret-for-watch-tokens";
+
+const {
+  cancelDashboardAgentWatch,
+  createDashboardAgentWatch,
+  deleteChatWithWatches,
+  listActiveWatchesForChats,
+} = await import("~/services/dashboardAgentWatches.server");
+const { sweepDashboardAgentWatches, WATCH_DELIVERY_GRACE_MS, WATCH_EXPIRY_GRACE_MS } =
+  await import("~/services/dashboardAgentWatchSweep.server");
+
+const harness = new DashboardAgentWatchesTestHarness(ctx, createDashboardAgentWatch);
+const boot = harness.boot.bind(harness);
+const seed = harness.seed.bind(harness);
+const authenticated = harness.authenticated.bind(harness);
+const seedChat = harness.seedChat.bind(harness);
+const runRow = harness.runRow.bind(harness);
+const fakeCheckDeps = harness.fakeCheckDeps.bind(harness);
+const create = harness.create.bind(harness);
+const storedMessages = harness.storedMessages.bind(harness);
+
+beforeEach(() => harness.reset());
+afterEach(() => harness.close());
+
+describe("the chat cascade and the list view", () => {
+  postgresTest(
+    "deleting a chat soft-deletes it and cancels its active watches in one call",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "cascade");
+      await seedChat(seeded, "chat_1");
+      await seedChat(seeded, "chat_2");
+
+      const mine = await create({ seeded, chatId: "chat_1" });
+      const theirs = await create({ seeded, chatId: "chat_2" });
+      expect(mine.ok && theirs.ok).toBe(true);
+      if (!mine.ok || !theirs.ok) return;
+
+      expect(
+        await deleteChatWithWatches({
+          chatId: "chat_1",
+          userId: seeded.user.id,
+          organizationId: seeded.organization.id,
+        })
+      ).toEqual({
+        deleted: true,
+        cancelledWatches: 1,
+      });
+
+      expect(
+        await chatExists(ctx.agentDb, {
+          chatId: "chat_1",
+          userId: seeded.user.id,
+          organizationId: seeded.organization.id,
+        })
+      ).toBe(false);
+      expect(await getWatch(ctx.agentDb, { id: mine.watchId })).toMatchObject({
+        status: "cancelled",
+        cancelReason: "chat_deleted",
+        deliveryStatus: "not_required",
+      });
+      expect(await getWatch(ctx.agentDb, { id: theirs.watchId })).toMatchObject({
+        status: "active",
+      });
+    }
+  );
+
+  postgresTest(
+    "a user's own cancel leaves one neutral line in the chat, and only one",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "usercancel");
+      await seedChat(seeded, "chat_1");
+
+      const created = await create({ seeded, chatId: "chat_1" });
+      expect(created.ok).toBe(true);
+      if (!created.ok) return;
+
+      const cancel = () =>
+        cancelDashboardAgentWatch({
+          watchId: created.watchId,
+          userId: seeded.user.id,
+          organizationId: seeded.organization.id,
+        });
+
+      expect(await cancel()).toMatchObject({
+        cancelled: true,
+        messages: [
+          {
+            id: `watch-cancelled:${created.watchId}`,
+            role: "assistant",
+            parts: [{ type: "text", text: "Stopped watching run run_1." }],
+          },
+        ],
+      });
+      expect(await getWatch(ctx.agentDb, { id: created.watchId })).toMatchObject({
+        status: "cancelled",
+        cancelReason: "user",
+        deliveryStatus: "not_required",
+      });
+      expect(await storedMessages(seeded, "chat_1")).toMatchObject([
+        { id: `watch-cancelled:${created.watchId}`, role: "assistant" },
+      ]);
+
+      // The row is no longer active, so the second cancel writes nothing at all.
+      expect(await cancel()).toEqual({ cancelled: false, messages: [] });
+      expect(await storedMessages(seeded, "chat_1")).toHaveLength(1);
+    }
+  );
+
+  postgresTest(
+    "a chat delete cancels its watches without a line in the chat",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "silentcancel");
+      await seedChat(seeded, "chat_1");
+
+      const created = await create({ seeded, chatId: "chat_1" });
+      expect(created.ok).toBe(true);
+
+      await deleteChatWithWatches({
+        chatId: "chat_1",
+        userId: seeded.user.id,
+        organizationId: seeded.organization.id,
+      });
+
+      const rows = await ctx.prisma.$queryRawUnsafe<{ message_id: string }[]>(
+        `select message_id from trigger_dashboard_agent.chat_messages where chat_id = 'chat_1'`
+      );
+      expect(rows).toEqual([]);
+    }
+  );
+
+  postgresTest(
+    "aggregates active watches per chat in one query",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "chips");
+      await seedChat(seeded, "chat_1");
+      await seedChat(seeded, "chat_2");
+
+      const a = await create({ seeded, chatId: "chat_1", spec: { ...RUN_START, runId: "run_1" } });
+      const b = await create({ seeded, chatId: "chat_1", spec: { ...RUN_START, runId: "run_2" } });
+      const c = await create({ seeded, chatId: "chat_2" });
+      expect(a.ok && b.ok && c.ok).toBe(true);
+
+      const byChat = await listActiveWatchesForChats({
+        chatIds: ["chat_1", "chat_2", "chat_missing"],
+        organizationId: seeded.organization.id,
+        userId: seeded.user.id,
+      });
+      expect(byChat.chat_1).toHaveLength(2);
+      expect(byChat.chat_2).toHaveLength(1);
+      expect(byChat.chat_missing).toBeUndefined();
+      expect(byChat.chat_2![0]).toMatchObject({
+        identity: "run_start:run_1",
+        status: "active",
+        kind: "run_start",
+        note: RUN_START.note,
+      });
+
+      if (a.ok) await cancelWatch(ctx.agentDb, { id: a.watchId, reason: "user" });
+      if (b.ok) await cancelWatch(ctx.agentDb, { id: b.watchId, reason: "user" });
+      expect(
+        (
+          await listActiveWatchesForChats({
+            chatIds: ["chat_1"],
+            organizationId: seeded.organization.id,
+            userId: seeded.user.id,
+          })
+        ).chat_1
+      ).toBeUndefined();
+    }
+  );
+
+  postgresTest("returns nothing for an empty chat list", async ({ prisma, postgresContainer }) => {
+    await boot(prisma, postgresContainer.getConnectionUri());
+    expect(
+      await listActiveWatchesForChats({ chatIds: [], organizationId: "org_x", userId: "user_x" })
+    ).toEqual({});
+  });
+});
+
+describe("unread watch wakes", () => {
+  postgresTest(
+    "only signals a wake once its delivery landed",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "unread");
+      await seedChat(seeded, "chat_1");
+
+      const created = await create({ seeded, chatId: "chat_1" });
+      expect(created.ok).toBe(true);
+      if (!created.ok) return;
+
+      const scope = { organizationId: seeded.organization.id, userId: seeded.user.id };
+      const recent = { ...scope, deliveredAfter: new Date(Date.now() - 15 * 60 * 1000) };
+
+      if (!created.watching) throw new Error("expected a watch");
+      await transitionWatchCondition(ctx.agentDb, {
+        id: created.watchId,
+        resolution: "condition_met",
+      });
+      expect(await countUnreadWatchWakes(ctx.agentDb, scope)).toBe(0);
+      expect(await listRecentWatchWakes(ctx.agentDb, recent)).toEqual([]);
+      expect(await listChatIdsWithUnreadWakes(ctx.agentDb, scope)).toEqual(new Set());
+
+      await markWatchDelivered(ctx.agentDb, { id: created.watchId });
+      expect(await countUnreadWatchWakes(ctx.agentDb, scope)).toBe(1);
+      expect(await listRecentWatchWakes(ctx.agentDb, recent)).toMatchObject([
+        { watchId: created.watchId, chatId: "chat_1", outcome: "fired", unread: true },
+      ]);
+      expect(await listChatIdsWithUnreadWakes(ctx.agentDb, scope)).toEqual(new Set(["chat_1"]));
+
+      // The poll's single query answers both halves the same way.
+      expect(await readWatchWakeFeed(ctx.agentDb, recent)).toMatchObject({
+        unreadWakes: 1,
+        wakes: [{ watchId: created.watchId, chatId: "chat_1", outcome: "fired", unread: true }],
+      });
+
+      // An unread wake from before the window still counts, but isn't narrated again.
+      expect(
+        await readWatchWakeFeed(ctx.agentDb, {
+          ...scope,
+          deliveredAfter: new Date(Date.now() + 60_000),
+        })
+      ).toMatchObject({ unreadWakes: 1, wakes: [] });
+    }
+  );
+});
+
+describe("the watch sweep", () => {
+  async function overdueWatch(seeded: Seeded, chatId = "chat_1") {
+    const created = await create({ seeded, chatId });
+    if (!created.ok) throw new Error("the watch wasn't created");
+    await ctx.prisma.$executeRawUnsafe(
+      `update trigger_dashboard_agent.watches set expires_at = now() - interval '1 hour' where id = $1`,
+      created.watchId
+    );
+    return created.watchId;
+  }
+
+  function sweepDeps(args: {
+    seeded: Seeded;
+    checkDeps?: Partial<WatchCheckDeps>;
+    revoked?: boolean;
+    now?: Date;
+    failDelivery?: boolean;
+    delivered: string[];
+  }) {
+    return {
+      now: () => args.now ?? new Date(),
+      checkDeps: () => fakeCheckDeps(args.checkDeps),
+      authorize: async () =>
+        args.revoked
+          ? ({ ok: false, reason: "access_revoked" } as const)
+          : ({ ok: true, environment: authenticated(args.seeded) } as const),
+      deliver: async (watch: Watch) => {
+        if (args.failDelivery) throw new Error("the delivery couldn't be scheduled");
+        args.delivered.push(watch.id);
+      },
+      configured: () => true,
+    };
+  }
+
+  postgresTest(
+    "runs the final check on an overdue watch and fires it at the buzzer",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "sweep");
+      await seedChat(seeded);
+      const watchId = await overdueWatch(seeded);
+
+      const delivered: string[] = [];
+      const result = await sweepDashboardAgentWatches(
+        sweepDeps({
+          seeded,
+          delivered,
+          checkDeps: {
+            readRun: async () => runRow({ status: "EXECUTING", startedAt: new Date() }),
+          },
+        })
+      );
+
+      expect(result).toMatchObject({ overdue: 1, fired: 1, expired: 0, cancelled: 0, failed: 0 });
+      expect(await getWatch(ctx.agentDb, { id: watchId })).toMatchObject({
+        status: "fired",
+        deliveryStatus: "pending",
+      });
+      expect(delivered).toEqual([watchId]);
+    }
+  );
+
+  postgresTest(
+    "expires an overdue watch the check says hasn't happened, as verified",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "sweep");
+      await seedChat(seeded);
+      const watchId = await overdueWatch(seeded);
+
+      const delivered: string[] = [];
+      const result = await sweepDashboardAgentWatches(sweepDeps({ seeded, delivered }));
+
+      expect(result).toMatchObject({ overdue: 1, expired: 1, failed: 0 });
+      const row = await getWatch(ctx.agentDb, { id: watchId });
+      expect(row).toMatchObject({ status: "expired", deliveryStatus: "pending" });
+      expect(row?.lastResult).toMatchObject({ verified: true, reason: "not_met_by_expiry" });
+      expect(delivered).toEqual([watchId]);
+    }
+  );
+
+  postgresTest(
+    "cancels an overdue watch whose user lost access, and never wakes the chat",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "sweep");
+      await seedChat(seeded);
+      const watchId = await overdueWatch(seeded);
+
+      const delivered: string[] = [];
+      const result = await sweepDashboardAgentWatches(
+        sweepDeps({ seeded, delivered, revoked: true })
+      );
+
+      expect(result).toMatchObject({ overdue: 1, cancelled: 1, expired: 0, fired: 0, failed: 0 });
+      expect(await getWatch(ctx.agentDb, { id: watchId })).toMatchObject({
+        status: "cancelled",
+        cancelReason: "access_revoked",
+        deliveryStatus: "not_required",
+      });
+      expect(delivered).toEqual([]);
+    }
+  );
+
+  postgresTest(
+    "leaves a watch that is still inside its deadline alone",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "sweep");
+      await seedChat(seeded);
+      const created = await create({ seeded });
+      expect(created.ok).toBe(true);
+
+      const delivered: string[] = [];
+      const result = await sweepDashboardAgentWatches(sweepDeps({ seeded, delivered }));
+
+      expect(result).toMatchObject({ overdue: 0, undelivered: 0 });
+      expect(delivered).toEqual([]);
+    }
+  );
+
+  postgresTest(
+    "recovers a wake the delivery lost, through the real query, exactly once",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "sweep");
+      await seedChat(seeded);
+      const watchId = await overdueWatch(seeded);
+
+      const delivered: string[] = [];
+      await expect(
+        sweepDashboardAgentWatches(sweepDeps({ seeded, delivered, failDelivery: true }))
+      ).rejects.toThrow(/failed on 1 watches/);
+      expect(await getWatch(ctx.agentDb, { id: watchId })).toMatchObject({
+        status: "expired",
+        deliveryStatus: "pending",
+      });
+      expect(delivered).toEqual([]);
+
+      const later = new Date(Date.now() + WATCH_DELIVERY_GRACE_MS + 60_000);
+      const second = await sweepDashboardAgentWatches(sweepDeps({ seeded, delivered, now: later }));
+      expect(second).toMatchObject({ undelivered: 1, redelivered: 1, failed: 0 });
+      expect(delivered).toEqual([watchId]);
+
+      await markWatchDelivered(ctx.agentDb, { id: watchId });
+      const third = await sweepDashboardAgentWatches(sweepDeps({ seeded, delivered, now: later }));
+      expect(third).toMatchObject({ undelivered: 0, redelivered: 0 });
+      expect(delivered).toEqual([watchId]);
+    }
+  );
+
+  postgresTest(
+    "a deliverer that died mid-delivery is recovered, but a fresh claim is left alone",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "sweep");
+      await seedChat(seeded);
+      const watchId = await overdueWatch(seeded);
+
+      await sweepDashboardAgentWatches(sweepDeps({ seeded, delivered: [] }));
+
+      await ctx.prisma.$executeRawUnsafe(
+        `update trigger_dashboard_agent.watches
+         set delivery_status = 'delivering',
+             delivery_claimed_at = now(),
+             last_checked_at = now() - interval '1 hour'
+         where id = $1`,
+        watchId
+      );
+      expect(await sweepDashboardAgentWatches(sweepDeps({ seeded, delivered: [] }))).toMatchObject({
+        undelivered: 0,
+      });
+
+      await ctx.prisma.$executeRawUnsafe(
+        `update trigger_dashboard_agent.watches
+         set delivery_claimed_at = now() - interval '1 hour' where id = $1`,
+        watchId
+      );
+      const recovered: string[] = [];
+      expect(
+        await sweepDashboardAgentWatches(sweepDeps({ seeded, delivered: recovered }))
+      ).toMatchObject({ undelivered: 1, redelivered: 1, failed: 0 });
+      expect(recovered).toEqual([watchId]);
+    }
+  );
+
+  postgresTest(
+    "leaves nothing owed for a request the immediate check already answered",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "sweep");
+      await seedChat(seeded);
+
+      const created = await create({
+        seeded,
+        checkDeps: { readRun: async () => runRow({ status: "EXECUTING", startedAt: new Date() }) },
+      });
+      expect(created.ok).toBe(true);
+      if (!created.ok || created.watching) throw new Error("expected a one-shot result");
+
+      const delivered: string[] = [];
+      const later = new Date(Date.now() + WATCH_DELIVERY_GRACE_MS + 60_000);
+      const result = await sweepDashboardAgentWatches(sweepDeps({ seeded, delivered, now: later }));
+
+      expect(result).toMatchObject({ overdue: 0, undelivered: 0, redelivered: 0 });
+      expect(delivered).toEqual([]);
+    }
+  );
+
+  postgresTest(
+    "finalizes overdue watches even with no agent to deliver to, and delivers once it's back",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "sweep");
+      await seedChat(seeded);
+      const watchId = await overdueWatch(seeded);
+
+      const delivered: string[] = [];
+      const unconfigured = await sweepDashboardAgentWatches({
+        ...sweepDeps({ seeded, delivered }),
+        configured: () => false,
+      });
+
+      expect(unconfigured).toMatchObject({
+        overdue: 1,
+        expired: 1,
+        deliveryDeferred: 1,
+        undelivered: 0,
+        redelivered: 0,
+        failed: 0,
+      });
+      expect(delivered).toEqual([]);
+      expect(await getWatch(ctx.agentDb, { id: watchId })).toMatchObject({
+        status: "expired",
+        deliveryStatus: "pending",
+      });
+
+      const later = new Date(Date.now() + WATCH_DELIVERY_GRACE_MS + 60_000);
+      const restored = await sweepDashboardAgentWatches(
+        sweepDeps({ seeded, delivered, now: later })
+      );
+      expect(restored).toMatchObject({ undelivered: 1, redelivered: 1, failed: 0 });
+      expect(delivered).toEqual([watchId]);
+    }
+  );
+
+  postgresTest(
+    "the expiry grace keeps the sweep off a watch the tick chain is still finishing",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "sweep");
+      await seedChat(seeded);
+      const created = await create({ seeded });
+      expect(created.ok).toBe(true);
+      if (!created.ok) return;
+
+      // A second past the deadline, so the chain's own final check owns this window.
+      await ctx.prisma.$executeRawUnsafe(
+        `update trigger_dashboard_agent.watches set expires_at = now() - interval '1 second' where id = $1`,
+        created.watchId
+      );
+      const delivered: string[] = [];
+      expect(await sweepDashboardAgentWatches(sweepDeps({ seeded, delivered }))).toMatchObject({
+        overdue: 0,
+      });
+
+      const later = new Date(Date.now() + WATCH_EXPIRY_GRACE_MS + 60_000);
+      expect(
+        await sweepDashboardAgentWatches(sweepDeps({ seeded, delivered, now: later }))
+      ).toMatchObject({ overdue: 1, expired: 1 });
+    }
+  );
+});
+
+describe("the tick claim", () => {
+  postgresTest(
+    "claiming a generation is not an observation: only a recorded check stamps one",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "claim");
+      await seedChat(seeded);
+      const created = await create({ seeded });
+      expect(created.ok).toBe(true);
+      if (!created.ok) return;
+
+      const claimed = await claimWatchTick(ctx.agentDb, { id: created.watchId, generation: 1 });
+      expect(claimed).toMatchObject({ tickCount: 1, lastCheckedAt: null, lastResult: null });
+
+      await recordWatchCheck(ctx.agentDb, { id: created.watchId, lastResult: { pending: 4 } });
+      const row = await getWatch(ctx.agentDb, { id: created.watchId });
+      expect(row?.lastCheckedAt).toBeInstanceOf(Date);
+      expect(row?.lastResult).toMatchObject({ pending: 4 });
+      expect(row?.tickCount).toBe(1);
+    }
+  );
+});
+
+// The delivery claim's fencing token: a hung deliverer is taken over, so an unfenced release or mark would touch the new owner's claim.
+describe("the delivery claim", () => {
+  async function firedWatch(seeded: Seeded) {
+    const created = await create({ seeded });
+    expect(created.ok).toBe(true);
+    if (!created.ok) throw new Error("the watch wasn't created");
+    const transitioned = await transitionWatchCondition(ctx.agentDb, {
+      id: created.watchId,
+      status: "fired",
+      lastResult: { result: "satisfied", facts: { verified: true } },
+    });
+    expect(transitioned).toMatchObject({ deliveryStatus: "pending" });
+    return created.watchId;
+  }
+
+  function staleBefore() {
+    return new Date(Date.now() - WATCH_DELIVERY_CLAIM_STALE_MS);
+  }
+
+  async function ageClaim(watchId: string) {
+    await ctx.prisma.$executeRawUnsafe(
+      `update trigger_dashboard_agent.watches
+       set delivery_claimed_at = now() - interval '1 hour' where id = $1`,
+      watchId
+    );
+  }
+
+  postgresTest(
+    "a stale takeover makes the old owner's release a no-op, and the new owner delivers once",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "claim-fence");
+      await seedChat(seeded);
+      const watchId = await firedWatch(seeded);
+
+      const a = await claimWatchDelivery(ctx.agentDb, { id: watchId, staleBefore: staleBefore() });
+      expect(a).not.toBeNull();
+      if (!a) return;
+
+      await ageClaim(watchId);
+      const b = await claimWatchDelivery(ctx.agentDb, { id: watchId, staleBefore: staleBefore() });
+      expect(b).not.toBeNull();
+      if (!b) return;
+      expect(b.claimId).not.toBe(a.claimId);
+
+      expect(
+        await releaseWatchDelivery(ctx.agentDb, { id: watchId, claimId: a.claimId })
+      ).toBeNull();
+      expect(await getWatch(ctx.agentDb, { id: watchId })).toMatchObject({
+        deliveryStatus: "delivering",
+        deliveryClaimId: b.claimId,
+      });
+
+      expect(
+        await claimWatchDelivery(ctx.agentDb, { id: watchId, staleBefore: staleBefore() })
+      ).toBeNull();
+
+      expect(
+        await markWatchDelivered(ctx.agentDb, { id: watchId, claimId: b.claimId })
+      ).toMatchObject({ deliveryStatus: "delivered" });
+      expect(await markWatchDelivered(ctx.agentDb, { id: watchId, claimId: b.claimId })).toBeNull();
+    }
+  );
+
+  postgresTest(
+    "a late delivered-mark from the old owner completes nothing",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "claim-late");
+      await seedChat(seeded);
+      const watchId = await firedWatch(seeded);
+
+      const a = await claimWatchDelivery(ctx.agentDb, { id: watchId, staleBefore: staleBefore() });
+      expect(a).not.toBeNull();
+      if (!a) return;
+      await ageClaim(watchId);
+      const b = await claimWatchDelivery(ctx.agentDb, { id: watchId, staleBefore: staleBefore() });
+      expect(b).not.toBeNull();
+      if (!b) return;
+
+      expect(await markWatchDelivered(ctx.agentDb, { id: watchId, claimId: a.claimId })).toBeNull();
+      expect(await markWatchDelivered(ctx.agentDb, { id: watchId })).toBeNull();
+      expect(await getWatch(ctx.agentDb, { id: watchId })).toMatchObject({
+        deliveryStatus: "delivering",
+        deliveredAt: null,
+      });
+
+      expect(
+        await markWatchDelivered(ctx.agentDb, { id: watchId, claimId: b.claimId })
+      ).toMatchObject({ deliveryStatus: "delivered" });
+    }
+  );
+
+  postgresTest(
+    "the inline path marks a pending delivery without a claim",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "claim-inline");
+      await seedChat(seeded);
+      const watchId = await firedWatch(seeded);
+
+      expect(await markWatchDelivered(ctx.agentDb, { id: watchId })).toMatchObject({
+        deliveryStatus: "delivered",
+      });
+      expect(await markWatchDelivered(ctx.agentDb, { id: watchId })).toBeNull();
+      expect(
+        await claimWatchDelivery(ctx.agentDb, { id: watchId, staleBefore: staleBefore() })
+      ).toBeNull();
+    }
+  );
+});
+
+describe("deleting a chat while a watch is being created", () => {
+  postgresTest("holds in both orders", async ({ prisma, postgresContainer }) => {
+    await boot(prisma, postgresContainer.getConnectionUri());
+    const seeded = await seed(prisma, "race");
+
+    for (const deleteFirst of [true, false]) {
+      const chatId = `chat_${deleteFirst ? "del" : "add"}`;
+      await seedChat(seeded, chatId);
+
+      const creating = () => create({ seeded, chatId });
+      const deleting = () =>
+        deleteChatWithWatches({
+          chatId,
+          userId: seeded.user.id,
+          organizationId: seeded.organization.id,
+        });
+      const [a, b] = deleteFirst
+        ? await Promise.all([deleting(), creating()])
+        : await Promise.all([creating(), deleting()]);
+      expect(a).toBeDefined();
+      expect(b).toBeDefined();
+
+      expect(await listActiveWatchesForChat(ctx.agentDb, { chatId })).toEqual([]);
+      expect(
+        await chatExists(ctx.agentDb, {
+          chatId,
+          userId: seeded.user.id,
+          organizationId: seeded.organization.id,
+        })
+      ).toBe(false);
+    }
+  });
+
+  postgresTest(
+    "refuses a create against an already-deleted chat",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "race");
+      await seedChat(seeded);
+      await deleteChatWithWatches({
+        chatId: "chat_1",
+        userId: seeded.user.id,
+        organizationId: seeded.organization.id,
+      });
+
+      expect(await create({ seeded })).toMatchObject({ ok: false, code: "chat_not_found" });
+      expect(await listActiveWatchesForChat(ctx.agentDb, { chatId: "chat_1" })).toEqual([]);
+    }
+  );
+});
+
+describe("appendChatMessageOnce", () => {
+  postgresTest(
+    "appends in order without rewriting the transcript",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "append");
+      await seedChat(seeded);
+
+      const first = { id: "watch-card:watch_1", role: "assistant", parts: [] };
+      const second = { id: "watch-card:watch_2", role: "assistant", parts: [] };
+
+      expect(
+        await appendChatMessageOnce(ctx.agentDb, {
+          chatId: "chat_1",
+          userId: seeded.user.id,
+          organizationId: seeded.organization.id,
+          message: first,
+        })
+      ).toBe(true);
+      await appendChatMessageOnce(ctx.agentDb, {
+        chatId: "chat_1",
+        userId: seeded.user.id,
+        organizationId: seeded.organization.id,
+        message: second,
+      });
+
+      const messages = await getChatMessages(ctx.agentDb, {
+        chatId: "chat_1",
+        userId: seeded.user.id,
+        organizationId: seeded.organization.id,
+      });
+      expect(messages).toEqual([first, second]);
+    }
+  );
+
+  postgresTest(
+    "appends nothing for a chat the caller doesn't own",
+    async ({ prisma, postgresContainer }) => {
+      await boot(prisma, postgresContainer.getConnectionUri());
+      const seeded = await seed(prisma, "append-owner");
+      await seedChat(seeded);
+
+      expect(
+        await appendChatMessageOnce(ctx.agentDb, {
+          chatId: "chat_1",
+          userId: "user_someone_else",
+          organizationId: seeded.organization.id,
+          message: { id: "watch-card:watch_1", role: "assistant", parts: [] },
+        })
+      ).toBe(false);
+
+      const messages = await getChatMessages(ctx.agentDb, {
+        chatId: "chat_1",
+        userId: seeded.user.id,
+        organizationId: seeded.organization.id,
+      });
+      expect(messages).toEqual([]);
+    }
+  );
+});
